+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+unreleased
+----------
+
+Feature:
+- Add `redis_client` option to provide a custom Redis client.
+
 v1.0
 ----